api_yamdb/requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
asgiref==3.2.10
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.11
colorama==0.4.4
Django==2.2.16
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.0.4
flake8==4.0.1
idna==3.3
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.3
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.1.0
pyparsing==3.0.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2020.1
requests==2.26.0
sqlparse==0.3.1
toml==0.10.2
urllib3==1.26.8
